Adjusted heights to be 100% instead of 100vh
Mobile browsers have this weird property that the height stays the same
when the address bar is shown or hidden. When the address bar is shown,
100% is the correct viewport size. This is smaller than 100vh! 100vh is
the correct viewport size when the address bar is hidden, at which 100%
would be too small to fill the full viewport.

I decided to make the page not scroll and thus always show the address
bar. Therefore I changed to using 100% instead of 100vh as height. If
a user manages to hide the address bar in some way, that should not be a
problem. The app might in that case just not fill the entire viewport.